# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zuula API
  version: 0.1.0
  description: |
    Uganda Fact-Guard (Zuula). Two surfaces share this document:

    - **Core API** (`/api/v1/*`) — used by the Zuula web app. Cookie session (or bearer JWT for
      non-browser core clients). Roles: `public`, `journalist`, `expert`, `admin`, gated per
      operation.
    - **Partner API** (`/v1/*`, served from `api.zuula.ug` in production) — `Authorization: Bearer
      <key>` with an API key, 100 requests/hour. This surface is documented for external partners
      on `/developers` and this contract matches that page exactly: same base URL shape, same
      four endpoints, same field names, same `Authorization: Bearer` header (not `X-API-Key`),
      same page-based pagination shape. See "Deviations from the P2 brief" in the P2 contract
      summary sent to Noah for the two places this contract intentionally differs from the
      brief's defaults in order to match the live `/developers` page and the working frontend
      code.

    All timestamps are ISO-8601 UTC. Requirement IDs (`FR-AUTH-01`, `§9.2`, …) refer to the CIT
    *System Development Plan & Functional Requirements Specification v1.0*.
  contact:
    name: Noah Prinzy
    url: https://github.com/Noah-Prinzy/zuula
servers:
  - url: https://zuula.ug
    description: Web app origin. Core API paths already include /api/v1.
  - url: https://api.zuula.ug
...
security:
  - sessionAuth: []
paths:
...
  /api/v1/me/api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: My partner API keys.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create a key. The full secret is returned once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/ApiScope" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiKey"
                  - type: object
                    required: [secret]
                    properties:
                      secret: { type: string, description: "Shown only in this response.", example: "zl_live_4f7a9c2e1b8d4a6f9c0e2b7a1d3f5e8c" }
        "400": { $ref: "#/components/responses/BadRequest" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/me/api-keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke a key.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked. }
        "404": { $ref: "#/components/responses/NotFound" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/fact-checks:
    get:
      operationId: searchFactChecks
      tags: [fact-checks]
      summary: Search and filter the Library (FR-SEARCH-01…04).
      security: []
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - name: verdict
          in: query
          description: Comma-separated `Verdict` values.
          schema: { type: string }
        - { name: category, in: query, schema: { type: string } }
        - { name: language, in: query, schema: { type: string } }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/ContentType" }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - name: sort
          in: query
          schema: { type: string, enum: [relevance, newest, most-rated], default: relevance }
        - { $ref: "#/components/parameters/Page" }
        - { name: perPage, in: query, schema: { type: integer, default: 12, minimum: 1, maximum: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/FactCheckSummary" }
...
  /api/v1/fact-checks/{id}:
    get:
      operationId: getFactCheck
      tags: [fact-checks]
      summary: Full report (FR-DETECT, FR-EXPLAIN, FR-REVIEW-04).
      security: []
      parameters:
        - { $ref: "#/components/parameters/FactCheckId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FactCheckReport" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/fact-checks/{id}/related:
    get:
      operationId: getRelatedFactChecks
      tags: [fact-checks]
      summary: Related reports by topic, source and wording overlap (FR-SEARCH-03).
      security: []
      parameters:
        - { $ref: "#/components/parameters/FactCheckId" }
        - name: limit
          in: query
          schema: { type: integer, default: 3, maximum: 10 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FactCheckSummary" }

  # ---------------------------------------------------------------------------------------
  # Ratings, CCS and comments (FR-RATE-01…10)
  # ---------------------------------------------------------------------------------------
...
  /v1/checks:
    servers: [{ url: https://api.zuula.ug }]
    post:
      operationId: partnerSubmitCheck
      tags: [partner]
      summary: Submit content for verification.
      security: [{ partnerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerCheckInput" }
          multipart/form-data:
            schema: { $ref: "#/components/schemas/PartnerCheckInput" }
      responses:
        "202":
          description: Queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionAccepted" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMedia" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/RateLimited" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/checks/{trackingId}:
    servers: [{ url: https://api.zuula.ug }]
    get:
      operationId: partnerGetCheckStatus
      tags: [partner]
      summary: Poll a check's status; `result` is present once completed.
      security: [{ partnerApiKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/TrackingId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerCheckStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/fact-checks/{id}:
    servers: [{ url: https://api.zuula.ug }]
    get:
      operationId: partnerGetFactCheck
      tags: [partner]
      summary: A published fact-check.
      security: [{ partnerApiKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/FactCheckId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FactCheckPublic" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/fact-checks:
    servers: [{ url: https://api.zuula.ug }]
    get:
      operationId: partnerSearchFactChecks
      tags: [partner]
      summary: Search published fact-checks.
      security: [{ partnerApiKey: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: verdict, in: query, schema: { $ref: "#/components/schemas/Verdict" } }
        - { name: category, in: query, schema: { type: string } }
        - { name: language, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { $ref: "#/components/parameters/Page" }
        - { name: perPage, in: query, schema: { type: integer, default: 10, maximum: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/FactCheckSearchItem" }
                  page: { type: integer }
                  perPage: { type: integer }
                  total: { type: integer }
        "429": { $ref: "#/components/responses/RateLimited" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ---------------------------------------------------------------------------------------
  # Inbound integration webhooks (FR-SUBMIT-04) — Step 4. Called by WhatsApp/Telegram's own
  # servers, not by apps/web; not part of the documented partner API either, so neither
  # sessionAuth nor partnerApiKey applies. Real signature/secret verification is P3 — see
  # app/adapters/{whatsapp,telegram}.py and docs/adr/0001-api-architecture.md.
  # ---------------------------------------------------------------------------------------
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: zuula_session
      description: "Core API. An opaque, server-side session token, set as this cookie on sign-in. Non-browser core clients may send the same token as `Authorization: Bearer <token>` instead (ADR 0002 §5)."
    partnerApiKey:
      type: http
      scheme: bearer
      bearerFormat: "zl_live_*"
      description: |
        Partner API. Matches `/developers` exactly: `Authorization: Bearer zl_live_...`
        (deliberately not the `X-API-Key` header the P2 brief first proposed — see the
        deviations note in `info.description`).
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
    TrackingId:
      name: trackingId
      in: path
      required: true
      schema: { type: string, pattern: "^ZL-[A-Z2-9]{4}-[A-Z2-9]{2}$" }
      example: ZL-7K3P-Q9
    FactCheckId:
      name: id
      in: path
      required: true
      schema: { type: string }
      example: fc-2026-0142
  responses:
    BadRequest:
      description: Bad request.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    Unauthorized:
      description: Missing or invalid credentials.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    NotFound:
      description: Not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
...
    PayloadTooLarge:
      description: File exceeds 50 MB.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    UnsupportedMedia:
      description: Unsupported file type.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    UnprocessableEntity:
      description: Validation failed.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    RateLimited:
      description: Rate limit exceeded (100 requests/hour per key).
      headers:
        X-RateLimit-Limit: { schema: { type: integer }, description: "Requests allowed per window." }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer }, description: Unix timestamp. }
        Retry-After: { schema: { type: integer }, description: Seconds. }
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
...
    Verdict:
      type: string
      enum: [authentic, likely-false, "false", ai-generated, unverifiable]
      description: FR-DETECT-01.
    ContentType:
      type: string
      enum: [text, url, image, audio, video]
    ClaimAssessment:
      type: string
      enum: ["false", misleading, unsupported, out-of-context, supported]
...
    ApiScope:
      type: string
      enum: [submit, read]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - file_too_large
                - unsupported_media
                - invalid_content
                - rate_limited
                - server_error
            message: { type: string }
            retryAfter: { type: integer, description: Seconds. Present for rate_limited. }
      example:
        error:
          code: rate_limited
          message: This key has used its 100 requests for this hour.
          retryAfter: 1260
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        perPage: { type: integer }
        total: { type: integer }

    # ---- Fact-check report ----
    Claim:
      type: object
      required: [id, start, end, assessment, reason, citationIds]
      properties:
        id: { type: string }
        start: { type: integer, description: Character offset into submittedText. }
        end: { type: integer, description: Exclusive. }
        assessment: { $ref: "#/components/schemas/ClaimAssessment" }
        reason: { type: string }
        citationIds: { type: array, items: { type: string } }
...
    CommunityScore:
      type: object
      description: Computed server-side (§9.1 weights, §9.2 thresholds). Not stored client-side.
      required: [ccs, weightedAccurate, weightedInaccurate, accurateCount, inaccurateCount, total, status]
      properties:
        ccs: { type: [number, "null"], minimum: 0, maximum: 100 }
        weightedAccurate: { type: number }
        weightedInaccurate: { type: number }
        accurateCount: { type: integer }
        inaccurateCount: { type: integer }
        total: { type: integer }
        status: { type: string, enum: [verified, standard, questioned, escalated, suspended] }
...
    FactCheckReport:
      description: Full report — core API only (includes submittedText, annotations and comments).
      type: object
      required:
        [
          id, trackingId, title, contentType, language, submittedText, verdict, confidence,
          summary, whatIsFalse, whatIsTrue, claims, citations, aiSignals, annotations,
          community, category, checkedAt, processingSeconds,
        ]
      properties:
        id: { type: string }
        trackingId: { type: string }
        title: { type: string }
        contentType: { $ref: "#/components/schemas/ContentType" }
        language: { type: string }
        submittedText: { type: string }
        sourceUrl: { type: string, format: uri }
        verdict: { $ref: "#/components/schemas/Verdict" }
        confidence: { type: integer, minimum: 0, maximum: 100, description: FR-DETECT-02. }
        summary: { type: string }
        whatIsFalse: { type: array, items: { type: string } }
        whatIsTrue: { type: array, items: { type: string } }
        claims: { type: array, items: { $ref: "#/components/schemas/Claim" } }
        citations: { type: array, items: { $ref: "#/components/schemas/Citation" } }
        aiSignals: { type: array, items: { $ref: "#/components/schemas/AISignal" } }
        annotations: { type: array, items: { $ref: "#/components/schemas/ExpertAnnotation" } }
        humanReview:
          oneOf:
            - { $ref: "#/components/schemas/HumanReview" }
            - type: "null"
        community: { $ref: "#/components/schemas/CommunityRating" }
        category: { type: string }
        checkedAt: { type: string, format: date-time }
        processingSeconds: { type: number }
    FactCheckSummary:
      description: List-item shape for /fact-checks search, the Home feed and related reports (core API).
      type: object
      required: [id, trackingId, title, contentType, language, verdict, confidence, summary, category, checkedAt]
      properties:
        id: { type: string }
        trackingId: { type: string }
        title: { type: string }
        contentType: { $ref: "#/components/schemas/ContentType" }
        language: { type: string }
        verdict: { $ref: "#/components/schemas/Verdict" }
        confidence: { type: integer }
        summary: { type: string }
        category: { type: string }
        checkedAt: { type: string, format: date-time }
        score: { $ref: "#/components/schemas/CommunityScore" }
    FactCheckPublic:
      description: |
        Partner-facing report shape (GET /v1/fact-checks/{id}). Matches the `factCheck()` helper
        in apps/web/components/developers/examples.ts exactly: no submittedText, no annotations,
        no rating comments — only aggregate community counts.
      type: object
      required:
        [
          id, trackingId, url, title, contentType, language, category, verdict, confidence,
          summary, whatIsFalse, whatIsTrue, claims, citations, aiSignals, community, humanReview,
          checkedAt, processingSeconds,
        ]
      properties:
        id: { type: string }
        trackingId: { type: string }
        url: { type: string, format: uri, example: "https://zuula.ug/fact-checks/fc-2026-0142" }
        title: { type: string }
        contentType: { $ref: "#/components/schemas/ContentType" }
        language: { type: string }
        category: { type: string }
        verdict: { $ref: "#/components/schemas/Verdict" }
        confidence: { type: integer }
        summary: { type: string }
        whatIsFalse: { type: array, items: { type: string } }
        whatIsTrue: { type: array, items: { type: string } }
        claims: { type: array, items: { $ref: "#/components/schemas/Claim" } }
        citations: { type: array, items: { $ref: "#/components/schemas/Citation" } }
        aiSignals: { type: array, items: { $ref: "#/components/schemas/AISignal" } }
        community:
          type: object
          required: [accurate, inaccurate]
          properties:
            accurate: { $ref: "#/components/schemas/RatingCounts" }
            inaccurate: { $ref: "#/components/schemas/RatingCounts" }
        humanReview:
          oneOf:
            - { $ref: "#/components/schemas/HumanReview" }
            - type: "null"
        checkedAt: { type: string, format: date-time }
        processingSeconds: { type: number }
    FactCheckSearchItem:
      description: Partner search result item (GET /v1/fact-checks) — matches SEARCH_RESPONSE in examples.ts.
      type: object
      required: [id, trackingId, url, title, verdict, confidence, summary, category, language, checkedAt]
      properties:
        id: { type: string }
        trackingId: { type: string }
        url: { type: string, format: uri }
        title: { type: string }
        verdict: { $ref: "#/components/schemas/Verdict" }
        confidence: { type: integer }
        summary: { type: string }
        category: { type: string }
        language: { type: string }
        checkedAt: { type: string, format: date-time }

    # ---- Submissions ----
...
    SubmissionAccepted:
      type: object
      required: [trackingId, status, estimatedSeconds, statusUrl]
      properties:
        trackingId: { type: string, example: ZL-7K3P-Q9 }
        status: { type: string, enum: [queued] }
        estimatedSeconds: { type: integer, example: 10 }
        statusUrl: { type: string, format: uri }
...
    ApiKey:
      type: object
      required: [id, name, prefix, scopes, createdAt, lastUsedAt]
      properties:
        id: { type: string }
        name: { type: string }
        prefix: { type: string, example: zl_live_4f7a }
        scopes: { type: array, items: { $ref: "#/components/schemas/ApiScope" } }
        createdAt: { type: string, format: date }
        lastUsedAt:
          oneOf:
            - { type: string, format: date-time }
            - type: "null"
...
    PartnerCheckInput:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [text, url, media, article] }
        content: { type: string }
        url: { type: string, format: uri }
        file: { type: string, format: binary }
        headline: { type: string }
        language: { type: string, description: "A LocaleCode, or `auto`." }
    PartnerCheckStatus:
      type: object
      required: [trackingId, status, submittedAt]
      properties:
        trackingId: { type: string }
        status: { type: string, enum: [queued, processing, completed, failed] }
        submittedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        result: { $ref: "#/components/schemas/FactCheckPublic" }
        error: { $ref: "#/components/schemas/ErrorEnvelope" }